More redirects for the relax HTML manual at http://www.nmr-relax.com/manual/index.html.
This is required as the latex2html program used to build the manual has been modified to not
excluded the 'common words' from the page titles (see docs/devel/latex2html/).  These redirects are
linked from outside or from the relax API documentation.

The redirects are for the new pages
http://www.nmr-relax.com/manual/The_full_NS_2_site_3D_CPMG_model.html,
http://www.nmr-relax.com/manual/The_new_protocol_in_the_GUI.html, and
http://www.nmr-relax.com/The_relax_data_model.html.