More redirects for the relax HTML manual at http://www.nmr-relax.com/manual/index.html.
This is required as the latex2html program used to build the manual has been modified to not
excluded the 'common words' from the page titles (see docs/devel/latex2html/).  These redirects are
linked from outside of http://www.nmr-relax.com, as identified by the Google Webmaster Tools.

The redirects are for the new pages
http://www.nmr-relax.com/manual/Comparison_dispersion_analysis_software.html,
http://www.nmr-relax.com/manual/Dispersion_script_mode_loading_data.html,
http://www.nmr-relax.com/manual/Keeping_branch_up_date_using_svnmerge_py.html,
http://www.nmr-relax.com/manual/Ri_prime_theta_Hessians.html,
http://www.nmr-relax.com/manual/clustered_relaxation_dispersion_analysis.html.