Replicated title fix for the HTML version of the manual.
The replicated titles were causing the sections with the same name to all point to a single HTML
page with the title name, resulting in a mess in the contents and in the flow of following the
manual sections sequentially.  The relax 4.0.0 documentation has therefore been especially updated
with these fixes, even though the release does not contain the title name changes.